R¤TUI | MASSIVE Refactor, splitting EVERYTHING into multiple smaller files, improves drastically codebase readability
